+++ CHANGELOG.md
@@ -17,6 +17,11 @@
 
 ## Unreleased
 
+#### 🐞 Fixes
+
+- Fixed an issue with `proto activate` where previously set environment variables would not be unset when changing directories.
+- Fixed an issue with `proto activate` where `[env]` variables would sometimes not be set.
+
 #### ⚙️ Internal
 
 - Updated Rust to v1.88.0.